Modularize source
The rollup setup supports this nicely already.
