+++ src/api.mjs
@@ -1,14 +1,4 @@
-var rfc6265Converter = {
-  read: function (value) {
-    return value.replace(/(%[\dA-F]{2})+/gi, decodeURIComponent)
-  },
-  write: function (value) {
-    return encodeURIComponent(value).replace(
-      /%(2[346BF]|3[ACDEF]|40|5[BDE]|60|7[BCD])/g,
-      decodeURIComponent
-    )
-  }
-}
+import rfc6265Converter from './rfc6265'
 
 function extend () {
   var result = {}
