+++ src/rfc6265.mjs
@@ -1,0 +1,11 @@
+export default {
+  read: function (value) {
+    return value.replace(/(%[\dA-F]{2})+/gi, decodeURIComponent)
+  },
+  write: function (value) {
+    return encodeURIComponent(value).replace(
+      /%(2[346BF]|3[ACDEF]|40|5[BDE]|60|7[BCD])/g,
+      decodeURIComponent
+    )
+  }
+}
